moved parsedgltf into own file
