+++ Runtime/Scripts/ParsedGltf.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 99b06f1b39ff43f4687b693cdfb943c4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
